+++ compatibility/step_definitions/cdata.ts
@@ -1,0 +1,9 @@
+import { Given } from "@badeball/cypress-cucumber-preprocessor";
+
+Given(
+  "I have {int} <![CDATA[cukes]]> in my belly",
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  function (cukeCount: number) {
+    // no-op
+  }
+);