+++ compatibility/step_definitions/minimal.ts
@@ -1,0 +1,6 @@
+import { Given } from "@badeball/cypress-cucumber-preprocessor";
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+Given("I have {int} cukes in my belly", function (cukeCount: number) {
+  // no-op
+});