+++ compatibility/step_definitions/stack-traces.ts
@@ -1,0 +1,5 @@
+import { When } from "@badeball/cypress-cucumber-preprocessor";
+
+When("a step throws an exception", function () {
+  throw new Error("BOOM");
+});